Operaciones CRUD

+++ app/Http/Controllers/Admin/ProductController.php
@@ -16,6 +16,13 @@ class ProductController extends Controller
         ]);
     }
 
+    public function show(Product $product)
+    {
+        return view('admin.products.show')->with([
+            'product' => $product,
+        ]);
+    }
+
     public function create()
     {
         return view('admin.products.create');
@@ -36,22 +43,23 @@ class ProductController extends Controller
             ->withSuccess("El producto {$product->title} se creó con éxito");
     }
 
-    public function show(Product $product)
-    {
-        return view('products.show')->with([
-            'product' => $product,
-        ]);
-    }
 
     public function edit(Product $product)
     {
-        return view('products.edit')->with([
+        return view('admin.products.edit')->with([
             'product' => $product,
         ]);
     }
 
-    public function udpate(Request $request, Product $product)
+    public function update(Request $request, Product $product)
     {
+        $rules = [
+            'title' => ['required', 'max:255'],
+            'price' => ['required', 'integer', 'min:1', 'max:500'],
+        ];
+
+        $request->validate($rules);
+
         $product->fill($request->all());
 
         $product->save();
@@ -62,6 +70,11 @@ class ProductController extends Controller
 
     public function destroy(Request $request, Product $product)
     {
+        if ($product->transactions()->count() > 0) {
+            return redirect()->back()->withErrors('No se puede eliminar un producto con 
+                transacciones');
+        }
+
         $product->delete();
 
         return redirect()->route('products.index')
